v3-fixed

+++ ft_printhex.c
@@ -6,13 +6,13 @@
 /*   By: jesssanc <jesssanc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/11 09:21:23 by jesssanc          #+#    #+#             */
-/*   Updated: 2024/12/12 09:02:11 by jesssanc         ###   ########.fr       */
+/*   Updated: 2024/12/12 09:33:02 by jesssanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_printhex_low(unsigned long num)
+int	ft_printhex_low(unsigned long long num)
 {
 	char	result[16];
 	int		i;
@@ -40,7 +40,7 @@ int	ft_printhex_low(unsigned long num)
 	return (count);
 }
 
-int	ft_printhex_up(unsigned long num)
+int	ft_printhex_up(unsigned long long num)
 {
 	char	result[16];
 	int		i;
